refactor: adjust folder structure for home page
