+++ pages/home/FAQSection.module.scss
@@ -1,0 +1,7 @@
+.questionBox {
+  margin: 24px 0;
+}
+
+.questionBox:last-child {
+  margin-bottom: 0;
+}
