+++ pages/home/Home.module.scss
@@ -105,14 +105,6 @@
   background-color: $gray-light;
 }
 
-.questionBox {
-  margin: 24px 0;
-}
-
-.questionBox:last-child {
-  margin-bottom: 0;
-}
-
 /* common */
 .button {
   margin-top: 40px;
